Support Rails 5.2
I think it makes sense in the README to say that the gem officially
supports only a specific range of Rails versions, but I would
be really happy if we could get rid of the upper-bound gem locking.
It makes it difficult for people to upgrade to the latest version of
Rails, and some people end up creating their own forks to get
around the this. There has been talk among some of the
Rails contributors to avoid gem locking when possible.